Move readers to separate modules
